provide alternative to storing passwords / login credentials  in notebooks
A persisted profile for the GIS can be created by giving the GIS and
it's authorization credentials and specifying a profile name. The
profile is stored in the users home directory in a config file named
.arcgisprofile. The profile is NOT ENCRYPTED and you need to take care
to protect the saved profile using operating system security or other
means. Once a profile has been saved, passing the profile parameter by
itself uses the authorization credentials saved in the configuration
file by that profile name.