updated to fix #404 and based on feedback from Steve Kopp
A few comments on the raster API…

Why did you make a patterns of SUM and SUM_ignorenodata two separate
items, along with numerous other data statistics,  instead of adding a
NoData Boolean to SUM?  This is the pattern we have used in the past
which I would have expected to see.

“Statistics” needs a more descriptive name, Focal_Statistics,
Neighborhood_Statistics…  there are dozens of places we do a variety of
statistics so it can’t be function name.