Interception of SSL error when the GIS uses a self-signed cert or no cert at all. (#509)
This change adds a try/except block prior to calling portalpy.Portal
with the submitted parameters. If the Exception raised has a type that
corresponds to an ssl.SSLError object, a RuntimeError is raised to
communicate the user of a way to bypass this error by submitting
"verify_cert=False" when instantiating the GIS object.