Weather data retrieval for ZikaAnalysis ML example. (#649)
* Interception of SSL error when the GIS uses a self-signed cert or no cert at all.

This change adds a try/except block prior to calling portalpy.Portal
with the submitted parameters. If the Exception raised has a type that
corresponds to an ssl.SSLError object, a RuntimeError is raised to
communicate the user of a way to bypass this error by submitting
"verify_cert=False" when instantiating the GIS object.

* Progress on weather data retrieval from wunderground api

The wunderground API surprisingly does not allow requests for historical
weekly summaries, like the ones provided in the weekly html in the
current notebook. The api limits requests to daily requests. It has been
observed that the api requests are significantly faster than the weekly
weather html scrape, but I still need to build out the logic to
summarize multiple daily weather records into a composite weekly summary
for the airport. More progress coming soon.

* Updates to weather scrape

* Finished weather data retrieval - weather_data.pkl and temp weather data provided