URL Scheme updates for apps  (#1608)
* future proofing params and adding enhancements

Changed some parameters to make them more future-proof. For instance, `webmap` parameter was better suited as `map_item` as it may refer to a Web Map or a Mobile Map Package depending on the app.

Added some enhancements, such as universal link support by providing a `url_type` parameter. This parameter will allow the user to create a `Web` link or an `App` link.

* move new parameters to the end

Moving all of the new parameters to the end

* rename `map_item` back to `webmap`