Enabled more color options for heatmaps and cleaned up documentation
Heatmaps can now take colorbrewer/palettable palettes through the spatial.plot() method, as well as use lists of rgb arrays or colormaps as a list of colorstops. In the case of a list of colormaps, uses the base color of each one as the colorstop. Additionally, fixed added things into the documentation and fixed a table in the spatial.plot() reference that wasn't showing.