Added better error handling, fixed tests
Added error handling on query, query_streaming, and search that will return an error if there is an issue with the openCypherQuery. Before, it just returned an empty list from the function, now informs users.

Additionally, fixed an issue with UUIDs that was causing test data to not get added properly, making the unit tests fail