allow logging in to Portal with read only mode
fixes an issue where making certain POST calls to a Portal will yield a server-side error.  The calls for portal status and user are more appropriate as GET calls and should work in all circumstances.

I left the legacy call to portal status/version a POST, as we will are not supporting versions that old at this point and I'm not sure when this feature was introduced.

Resolves #13625